+++ src/components/PostList/PostListTitle.js
@@ -1,4 +1,4 @@
-import { push } from "../utils/router.js";
+import { push } from "../../utils/router.js";
 
 export default function PostListTitle({ $target }) {
   const $postListTitle = document.createElement("div");
@@ -6,7 +6,10 @@ export default function PostListTitle({ $target }) {
   $postListTitle.className = "post-list-title post-list-block";
 
   this.render = () => {
-    $postListTitle.innerHTML = `최희라의 Notion`;
+    $postListTitle.innerHTML = `
+      <img src="/src/icons/notion-logo.png" alt="logo" width="25"  />
+      <span style="margin-left: 5px;">최희라의 Notion</span>
+    `;
   };
 
   $postListTitle.addEventListener("click", () => {
