+++ src/style/style.css
@@ -30,7 +30,7 @@ body {
 
 .post-list-title {
   font-weight: 1000;
-  margin-left: 10px;
+  margin-left: 5px;
   height: 50px;
   color: #37352f;
 }
